Add files via upload
This project aims to apply the different principles that I have learnt in the course "Data Analytics". And to predict the customer churn rate along with a detailed test report of each of the models we have implemented like- Decision tree & Random forest classifier. We have applied tree pruning and Random stratified Sampling methods to workaround with Bias & Variance.